+++ tests/ui/test_activationkey.py
@@ -15,7 +15,8 @@ from nose.plugins.attrib import attr
 from robottelo.common.constants import NOT_IMPLEMENTED, ENVIRONMENT
 from robottelo.common.decorators import bzbug
 from robottelo.common.helpers import generate_name, valid_names_list
-from robottelo.ui.factory import Session, make_org
+from robottelo.ui.factory import make_org
+from robottelo.ui.session import Session
 from robottelo.ui.locators import common_locators
 from tests.ui.baseui import BaseUI
