+++ tests/ui/test_gpgkey.py
@@ -19,8 +19,9 @@ from robottelo.common.helpers import (generate_name, get_data_file,
                                       read_data_file, valid_names_list,
                                       invalid_names_list, valid_data_list,
                                       generate_strings_list)
-from robottelo.ui.factory import Session, make_org
+from robottelo.ui.factory import make_org
 from robottelo.ui.locators import common_locators
+from robottelo.ui.session import Session
 from tests.ui.baseui import BaseUI
 
 REPO_URL = "http://inecas.fedorapeople.org/fakerepos/zoo2/"
