+++ tests/ui/test_repository.py
@@ -6,8 +6,9 @@ from ddt import data, ddt
 from nose.plugins.attrib import attr
 from robottelo.common.decorators import bzbug
 from robottelo.common.helpers import generate_name, generate_strings_list
-from robottelo.ui.factory import Session, make_org
+from robottelo.ui.factory import make_org
 from robottelo.ui.locators import common_locators
+from robottelo.ui.session import Session
 from tests.ui.baseui import BaseUI
